kernel: use (u)intptr_t to define (U)Int
This way, we don't rely on config.h for this definition

+++ src/saveload.c
@@ -221,7 +221,7 @@ void SaveUInt( UInt data )
     SaveUInt8(data);
 }
 
-UInt8 LoadUInt ( void )
+UInt LoadUInt ( void )
 {
     return LoadUInt8();
 }
